find a website that has errors
The P5 web editor has a built-in console — but it turns out that Chrome also has a console to display errors. Do some googling to find out how you can open the Chrome console. Then, try to find a website that has errors — visit a website, open the console, and see if there are any red-highlighted items. Take a screenshot and upload it into your repository.